+++ custom_widgets/whats_new_controller.cpp
@@ -7,9 +7,7 @@ YACReader::WhatsNewController::WhatsNewController() { }
 
 void YACReader::WhatsNewController::showWhatsNewIfNeeded(QWidget *fromParent)
 {
-    QSettings commonSettings(QStandardPaths::writableLocation(QStandardPaths::GenericDataLocation) +
-                                     "/YACReader/YACReaderCommon.ini",
-                             QSettings::IniFormat);
+    QSettings commonSettings(YACReader::getCommonSettingsFilePath(), QSettings::IniFormat);
 
     if (commonSettings.value("LAST_VERSION_INSTALLED").toString() != VERSION) {
         showWhatsNew(fromParent);